+++ doc/CHANGES.rst
@@ -54,6 +54,8 @@ CHANGELOG
   * pyodbc
 
 * System.Parameter.get do not raise if a default value is provided
+* Add ``--with-demo`` parameter while creating a database (anyblok_createdb) in
+  order to load demo data while installing bloks on that database.
 
 0.22.5 (2019-06-24)
 -------------------